--- address/src/main/kotlin/htnk128/kotlin/ddd/sample/address/Application.kt
@@ -1,11 +1,0 @@
-package htnk128.kotlin.ddd.sample.address
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class Application
-
-fun main(args: Array<String>) {
-    runApplication<Application>(*args)
-}
